% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_gdal.R
\docType{data}
\name{annotation_gdal}
\alias{annotation_gdal}
\alias{GeomGdal}
\title{Add background imagery}
\format{
An object of class \code{GeomGdal} (inherits from \code{Geom}, \code{ggproto}, \code{gg}) of length 5.
}
\usage{
annotation_gdal(
  dsn = c("osm", "virtualearth"),
  resample = "bilinear",
  interpolate = FALSE,
  data = NULL,
  mapping = NULL,
  alpha = 1
)

GeomGdal
}
\arguments{
\item{dsn}{The map source (currently 'osm' or 'virtualearth' are built-in - whatarelief streetmap, or imagery, otherwise use a GDAL DSN)}

\item{interpolate}{Passed to \code{\link[grid:grid.raster]{grid::rasterGrob()}}}

\item{data, mapping}{Specify data and mapping to use this geom with facets}

\item{alpha}{Use to make this layer semi-transparent}
}
\value{
A ggplot2 layer
}
\description{
Uses OpenStreetMap or VirtualEarth to add background imagery, or a custom source via 'dsn'.
If you are publishing
a map using these tiles, make sure to use the proper attribution
(e.g., "Copyright OpenStreetMap contributors" when using an
OpenStreetMap-based tile set).  Ditto for VirtualEarth or any dsn you use.
}
\examples{
\donttest{
library(ggplot2)
data(iw)

ggplot() +
  annotation_gdal() +
  geom_sf(data = sf::st_transform(iw, "EPSG:3857"), fill = NA, col = "grey50")

 ggplot() +
  annotation_gdal("virtualearth") +
  geom_sf(data = iw, fill = NA, col = "grey50")


pts <- do.call(cbind, maps::map(plot = F)[1:2])
pts <- pts[!is.na(pts[,1]), ]
pts <- pts[seq(1, nrow(pts), length.out = 8000), ]
sf <- sf::st_sf(geom = sf::st_sfc(sf::st_multipoint(pts), crs = "OGC:CRS84"))
ggplot() +
  annotation_gdal(dsn = "virtualearth") +
  geom_sf(data = sf::st_transform(sf, "+proj=laea +lon_0=147 +lat_0=-42"), fill = NA, col = "yellow", pch = ".")

pts2 <- pts[pts[,1] > 50 & pts[,1] < 120 & pts[,2] < -20 & pts[,2] > -45, ]
sf <- sf::st_sf(geom = sf::st_sfc(sf::st_multipoint(pts2), crs = "OGC:CRS84"))
ggplot() +
  annotation_gdal(dsn = "osm") +
  geom_sf(data = sf::st_transform(sf, "EPSG:3577"), fill = NA, col = "yellow", pch = ".")
pts3 <- pts[ pts[,2] < -20, ]
sf <- sf::st_sf(geom = sf::st_sfc(sf::st_multipoint(pts3), crs = "OGC:CRS84"))
wms_arcgis_mapserver_tms <-
"<GDAL_WMS><Service name=\"TMS\"><ServerUrl>http://services.arcgisonline.com/ArcGIS/rest/services/World_Street_Map/MapServer/tile/${z}/${y}/${x}</ServerUrl></Service><DataWindow><UpperLeftX>-20037508.34</UpperLeftX><UpperLeftY>20037508.34</UpperLeftY><LowerRightX>20037508.34</LowerRightX><LowerRightY>-20037508.34</LowerRightY><TileLevel>17</TileLevel><TileCountX>1</TileCountX><TileCountY>1</TileCountY><YOrigin>top</YOrigin></DataWindow><Projection>EPSG:900913</Projection><BlockSizeX>256</BlockSizeX><BlockSizeY>256</BlockSizeY><BandsCount>3</BandsCount><MaxConnections>10</MaxConnections><Cache /></GDAL_WMS>"
ggplot() +
  annotation_gdal(dsn = wms_arcgis_mapserver_tms, resample ="lanczos") +
  geom_sf(data = sf::st_transform(sf, "EPSG:3031"), fill = NA, col = "hotpink", pch = 19, cex = 0.2)

}

}
\keyword{datasets}
