% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggdal-package.R, R/ggdal.R
\docType{package}
\name{ggdal-package}
\alias{ggdal}
\alias{ggdal-package}
\alias{GeomFromGDAL}
\title{ggdal: GDAL Rendering for 'ggplot2'}
\format{
An object of class \code{GeomFromGDAL} (inherits from \code{Geom}, \code{ggproto}, \code{gg}) of length 5.
}
\usage{
GeomFromGDAL
}
\description{
A 'ggplot2' extension that enables image grobs from GDAL data source strings (DSN) in 'ggplot2'. Ruthlessly ripped off from 'ggpath'.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/hypertidy/ggdal}
  \item Report bugs at \url{https://github.com/hypertidy/ggdal/issues}
}

}
\author{
\strong{Maintainer}: Michael D. Sumner \email{mdsumner@gmail.com}

}
\keyword{datasets}
\keyword{internal}
