% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggdal-package.R
\docType{package}
\name{ggdal-package}
\alias{ggdal}
\alias{ggdal-package}
\title{ggdal: GDAL Rendering for 'ggplot2'}
\description{
A 'ggplot2' extension that enables image data from GDAL data sources (DSN).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/hypertidy/ggdal}
  \item Report bugs at \url{https://github.com/hypertidy/ggdal/issues}
}

}
\author{
\strong{Maintainer}: Michael D. Sumner \email{mdsumner@gmail.com}

Authors:
\itemize{
  \item Dewey Dunnington (wrote ggspatial::annotation_map_tile that was a key example for annotation_gdal to work from) [contributor]
}

}
\keyword{internal}
